Fix BL-3071 by putting all page-chooser dialog stuff together. Sheesh
