+++ src/BloomBrowserUI/webpack.config.js
@@ -28,7 +28,7 @@ module.exports = {
              editablePageBundle: './bookEdit/editablePage.ts',
              bookPreviewBundle: './bookPreview/bookPreview.ts',
              toolboxBundle: './bookEdit/toolbox/toolboxBootstrap.ts',
-             pageChooserBundle: './pageChooser/page-chooser.ts',
+             pageChooserBundle: ['./pageChooser/page-chooser.ts'],//https://github.com/webpack/webpack/issues/300
              pageThumbnailListBundle: './bookEdit/pageThumbnailList/pageThumbnailList.ts',
              testBundle: globule.find(["./**/*Spec.ts", "./**/*Spec.js", "!./node_modules/**"])//TODO this maybe slow if 1st it finds it all, then it excludes node_modules
            },
